Minor stuff, getting prepared to do a major update
